requirements.txt
openai~=0.28.1
flask~=3.0.0
python-dotenv~=1.0.0
